backend/pyproject.toml
[tool.poetry]
name = "news-intelligence-backend"
version = "0.1.0"
description = "Real-Time News Intelligence Engine Backend"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
celery = "^5.3.6"
redis = "^5.0.1"
feedparser = "^6.0.11"
newspaper3k = "^0.2.8"
lxml-html-clean = "^0.1.1"
spacy = "^3.7.2"
vaderSentiment = "^3.3.2"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
